+++ nuxt.config.ts
@@ -2,6 +2,7 @@
 export default defineNuxtConfig({
   compatibilityDate: '2025-07-15',
   devtools: { enabled: false },
+  ssr: false,
   modules: ['@pinia/nuxt', '@vant/nuxt', '@vueuse/nuxt'],
   css: ['~/assets/scss/main.scss'],
 });