+++ types/exercise.ts
@@ -1,7 +1,10 @@
+import { type EquipmentType } from './equipment-type';
+
 export interface Exercise {
   id: string;
   muscleGroupId: string;
   name: string;
   isCustom: boolean;
+  equipment?: EquipmentType;
   order?: number;
 }